fix; move make_options_body
