+++ tests/unit/io/test_asyncorereactor.py
@@ -28,8 +28,8 @@ import socket
 from socket import error as socket_error
 from cassandra.connection import ConnectionException, ProtocolError
 from cassandra.io.asyncorereactor import AsyncoreConnection
-from cassandra.protocol import (write_stringmultimap, write_int, write_string,
-                                SupportedMessage, ReadyMessage, ServerError)
+from cassandra.protocol import (write_int, write_string, SupportedMessage,
+                                ReadyMessage, ServerError)
 from cassandra.marshal import uint32_pack, int32_pack
 from tests import is_monkey_patched
 from tests.unit.io.utils import submit_and_wait_for_completion, TimerCallback, ReactorTestMixin
@@ -37,7 +37,8 @@ from tests.unit.io.utils import submit_and_wait_for_completion, TimerCallback, R
 
 class AsyncoreConnectionTest(unittest.TestCase, ReactorTestMixin):
 
-    connection_classs = AsyncoreConnection
+    connection_class = AsyncoreConnection
+    socket_attr_name = 'socket'
 
     @classmethod
     def setUpClass(cls):
@@ -61,14 +62,6 @@ class AsyncoreConnectionTest(unittest.TestCase, ReactorTestMixin):
     def setUp(self):
         if is_monkey_patched():
             raise unittest.SkipTest("Can't test asyncore with monkey patching")
-
-    def make_options_body(self):
-        options_buf = BytesIO()
-        write_stringmultimap(options_buf, {
-            'CQL_VERSION': ['3.0.1'],
-            'COMPRESSION': []
-        })
-        return options_buf.getvalue()
 
     def make_error_body(self, code, msg):
         buf = BytesIO()
